Rename fitters to tt
